pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "template-fastapi"
version = "0.1.0"
description = "Template para projetos FastAPI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.103.0",
    "sqlalchemy>=2.0.0",
    "alembic",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "uvicorn>=0.23.0",
    "pytest>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]